app/requirements.txt
flask>=2.0.0
openai
google-cloud-firestore>=2.11.0
requests>=2.28.0
beautifulsoup4>=4.11.0
gunicorn>=20.1.0
python-dotenv